fix: refactoring dashboard by tabs +  UI fixes
